+++ directdm/num/single_nucleon_form_factors.py
@@ -1,9 +1,8 @@
 #!/usr/bin/env python3
 
-from directdm.num.num_input import Num_input
-
+        
 class F1:
-    def __init__(self, quark, nucleon):
+    def __init__(self, quark, nucleon, input_dict):
         """ The nuclear form factor F1
         
         Return the nuclear form factor F1
@@ -17,6 +16,7 @@ class F1:
 
         self.quark = quark
         self.nucleon = nucleon
+        self.ip = input_dict
 
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
@@ -36,9 +36,20 @@ class F1:
             if self.quark == 's':
                 return 0
 
+    def first_deriv_zero_mom(self):
+        """ Return the value of the first derivative of the form factor
+            w.r.t. q^2 at zero momentum transfer (only strange quark) """
+
+        if self.nucleon == 'p':
+            if self.quark == 's':
+                return self.ip['rs2'] / 6
+        if self.nucleon == 'n':
+            if self.quark == 's':
+                return self.ip['rs2'] / 6
+
 
 class F2(object):
-    def __init__(self, quark, nucleon, input_dict=None):
+    def __init__(self, quark, nucleon, input_dict):
         """ The nuclear form factor F2
 
         Return the nuclear form factor F2
@@ -54,37 +65,30 @@ class F2(object):
         """
         self.quark = quark
         self.nucleon = nucleon
+        self.ip = input_dict
 
-        if input_dict is None:
-            self.input_dict = Num_input().input_parameters
-            # One should include a warning in case the dictionary
-            # does not contain all necessary keys
-        else:
-            self.input_dict = input_dict
 
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
 
-        ip = self.input_dict
-        
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return 2*ip['ap'] + ip['an'] + ip['F2sp']
+                return 2*self.ip['ap'] + self.ip['an'] + self.ip['F2sp']
             if self.quark == 'd':
-                return 2*ip['an'] + ip['ap'] + ip['F2sp']
+                return 2*self.ip['an'] + self.ip['ap'] + self.ip['F2sp']
             if self.quark == 's':
-                return ip['F2sp']
+                return self.ip['F2sp']
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return 2*ip['an'] + ip['ap'] + ip['F2sp']
+                return 2*self.ip['an'] + self.ip['ap'] + self.ip['F2sp']
             if self.quark == 'd':
-                return 2*ip['ap'] + ip['an'] + ip['F2sp']
+                return 2*self.ip['ap'] + self.ip['an'] + self.ip['F2sp']
             if self.quark == 's':
-                return ip['F2sp']
+                return self.ip['F2sp']
 
 
 class FA(object):
-    def __init__(self, quark, nucleon, input_dict=None):
+    def __init__(self, quark, nucleon, input_dict):
         """ The nuclear form factor FA at zero momentum transfer
 
         Return the nuclear form factor FA, evaluated at zero momentum transfer.
@@ -100,36 +104,30 @@ class FA(object):
         """
         self.quark = quark
         self.nucleon = nucleon
+        self.ip = input_dict
 
-        if input_dict is None:
-            self.input_dict = Num_input().input_parameters
-            # One should include a warning in case the dictionary
-            # does not contain all necessary keys
-        else:
-            self.input_dict = input_dict
 
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = self.input_dict
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return ip['Deltaup']
+                return self.ip['Deltaup']
             if self.quark == 'd':
-                return ip['Deltadp']
+                return self.ip['Deltadp']
             if self.quark == 's':
-                return ip['Deltas']
+                return self.ip['Deltas']
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return ip['Deltaun']
+                return self.ip['Deltaun']
             if self.quark == 'd':
-                return ip['Deltadn']
+                return self.ip['Deltadn']
             if self.quark == 's':
-                return ip['Deltas']
+                return self.ip['Deltas']
 
 
 class FPprimed(object):
-    def __init__(self, quark, nucleon, input_dict=None):
+    def __init__(self, quark, nucleon, input_dict):
         """ The nuclear form factor FPprimed
 
         Return the nuclear form factor FPprimed
@@ -145,34 +143,29 @@ class FPprimed(object):
         """
         self.quark = quark
         self.nucleon = nucleon
+        self.ip = input_dict
 
-        if input_dict is None:
-            self.input_dict = Num_input().input_parameters
-            # One should include a warning in case the dictionary
-            # does not contain all necessary keys
-        else:
-            self.input_dict = input_dict
 
     def value_pion_pole(self):
         """ Return the coefficient of the pion pole
 
         The pion pole is given, in terms of the spatial momentum q, by 1 / (q^2 + mpi0^2)
         """
-        ip = self.input_dict
-        self.mN = (ip['mproton'] + ip['mneutron'])/2
+
+        self.mN = (self.ip['mproton'] + self.ip['mneutron'])/2
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return self.mN**2 * 2 * ip['gA']
+                return self.mN**2 * 2 * self.ip['gA']
             if self.quark == 'd':
-                return - self.mN**2 * 2 * ip['gA']
+                return - self.mN**2 * 2 * self.ip['gA']
             if self.quark == 's':
                 return 0
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return - self.mN**2 * 2 * ip['gA']
+                return - self.mN**2 * 2 * self.ip['gA']
             if self.quark == 'd':
-                return self.mN**2 * 2 * ip['gA']
+                return self.mN**2 * 2 * self.ip['gA']
             if self.quark == 's':
                 return 0
 
@@ -181,27 +174,27 @@ class FPprimed(object):
 
         The eta pole is given, in terms of the spatial momentum q, by 1 / (q^2 + meta^2)
         """
-        ip = self.input_dict
-        self.mN = (ip['mproton'] + ip['mneutron'])/2
+
+        self.mN = (self.ip['mproton'] + self.ip['mneutron'])/2
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return self.mN**2 * 2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3
+                return self.mN**2 * 2 * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])/3
             if self.quark == 'd':
-                return self.mN**2 * 2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3
+                return self.mN**2 * 2 * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])/3
             if self.quark == 's':
-                return - self.mN**2 * 4 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3
+                return - self.mN**2 * 4 * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])/3
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return self.mN**2 * 2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3
+                return self.mN**2 * 2 * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])/3
             if self.quark == 'd':
-                return self.mN**2 * 2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3
+                return self.mN**2 * 2 * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])/3
             if self.quark == 's':
-                return - self.mN**2 * 4 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3
+                return - self.mN**2 * 4 * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])/3
 
 
 class FS(object):
-    def __init__(self, quark, nucleon, input_dict=None):
+    def __init__(self, quark, nucleon, input_dict):
         """ The nuclear form factor FS
 
         Return the nuclear form factor FS
@@ -217,36 +210,30 @@ class FS(object):
         """
         self.quark = quark
         self.nucleon = nucleon
+        self.ip = input_dict
 
-        if input_dict is None:
-            self.input_dict = Num_input().input_parameters
-            # One should include a warning in case the dictionary
-            # does not contain all necessary keys
-        else:
-            self.input_dict = input_dict
 
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = self.input_dict
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return ip['sigmaup']
+                return self.ip['sigmaup']
             if self.quark == 'd':
-                return ip['sigmadp']
+                return self.ip['sigmadp']
             if self.quark == 's':
-                return ip['sigmas']
+                return self.ip['sigmas']
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return ip['sigmaun']
+                return self.ip['sigmaun']
             if self.quark == 'd':
-                return ip['sigmadn']
+                return self.ip['sigmadn']
             if self.quark == 's':
-                return ip['sigmas']
+                return self.ip['sigmas']
 
 
 class FP(object):
-    def __init__(self, quark, nucleon, input_dict=None):
+    def __init__(self, quark, nucleon, input_dict):
         """ The nuclear form factor FP
 
         Return the nuclear form factor FP
@@ -262,34 +249,29 @@ class FP(object):
         """
         self.quark = quark
         self.nucleon = nucleon
+        self.ip = input_dict
 
-        if input_dict is None:
-            self.input_dict = Num_input().input_parameters
-            # One should include a warning in case the dictionary
-            # does not contain all necessary keys
-        else:
-            self.input_dict = input_dict
 
     def value_pion_pole(self):
         """ Return the coefficient of the pion pole
 
         The pion pole is given, in terms of the spatial momentum q, by 1 / (q^2 + mpi0^2)
         """
-        ip = self.input_dict
-        self.mN = (ip['mproton'] + ip['mneutron'])/2
+
+        self.mN = (self.ip['mproton'] + self.ip['mneutron'])/2
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return self.mN**2 * ip['gA'] * ip['B0mu'] / self.mN 
+                return self.mN**2 * self.ip['gA'] * self.ip['B0mu'] / self.mN 
             if self.quark == 'd':
-                return - self.mN**2 * ip['gA'] * ip['B0md'] / self.mN 
+                return - self.mN**2 * self.ip['gA'] * self.ip['B0md'] / self.mN 
             if self.quark == 's':
                 return 0
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return - self.mN**2 * ip['gA'] * ip['B0mu'] / self.mN 
+                return - self.mN**2 * self.ip['gA'] * self.ip['B0mu'] / self.mN 
             if self.quark == 'd':
-                return self.mN**2 * ip['gA'] * ip['B0md'] / self.mN 
+                return self.mN**2 * self.ip['gA'] * self.ip['B0md'] / self.mN 
             if self.quark == 's':
                 return 0
 
@@ -298,27 +280,27 @@ class FP(object):
 
         The eta pole is given, in terms of the spatial momentum q, by 1 / (q^2 + meta^2)
         """
-        ip = self.input_dict
-        self.mN = (ip['mproton'] + ip['mneutron'])/2
+
+        self.mN = (self.ip['mproton'] + self.ip['mneutron'])/2
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return self.mN**2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3/self.mN * ip['B0mu']
+                return self.mN**2 * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])/3/self.mN * self.ip['B0mu']
             if self.quark == 'd':
-                return self.mN**2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3/self.mN * ip['B0md']
+                return self.mN**2 * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])/3/self.mN * self.ip['B0md']
             if self.quark == 's':
-                return - 2 * self.mN**2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3/self.mN * ip['B0ms']
+                return - 2 * self.mN**2 * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])/3/self.mN * self.ip['B0ms']
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return self.mN**2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3/self.mN * ip['B0mu']
+                return self.mN**2 * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])/3/self.mN * self.ip['B0mu']
             if self.quark == 'd':
-                return self.mN**2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3/self.mN * ip['B0md']
+                return self.mN**2 * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])/3/self.mN * self.ip['B0md']
             if self.quark == 's':
-                return - 2 * self.mN**2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3/self.mN * ip['B0ms']
+                return - 2 * self.mN**2 * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])/3/self.mN * self.ip['B0ms']
 
 
 class FT0(object):
-    def __init__(self, quark, nucleon, input_dict=None):
+    def __init__(self, quark, nucleon, input_dict):
         """ The nuclear form factor FT0
 
         Return the nuclear form factor FT0
@@ -334,36 +316,30 @@ class FT0(object):
         """
         self.quark = quark
         self.nucleon = nucleon
+        self.ip = input_dict
 
-        if input_dict is None:
-            self.input_dict = Num_input().input_parameters
-            # One should include a warning in case the dictionary
-            # does not contain all necessary keys
-        else:
-            self.input_dict = input_dict
 
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = self.input_dict
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return ip['mu_at_2GeV'] * ip['gTu']
+                return self.ip['mu_at_2GeV'] * self.ip['gTu']
             if self.quark == 'd':
-                return ip['md_at_2GeV'] * ip['gTd']
+                return self.ip['md_at_2GeV'] * self.ip['gTd']
             if self.quark == 's':
-                return ip['ms_at_2GeV'] * ip['gTs']
+                return self.ip['ms_at_2GeV'] * self.ip['gTs']
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return ip['mu_at_2GeV'] * ip['gTd']
+                return self.ip['mu_at_2GeV'] * self.ip['gTd']
             if self.quark == 'd':
-                return ip['md_at_2GeV'] * ip['gTu']
+                return self.ip['md_at_2GeV'] * self.ip['gTu']
             if self.quark == 's':
-                return ip['ms_at_2GeV'] * ip['gTs']
+                return self.ip['ms_at_2GeV'] * self.ip['gTs']
 
 
 class FT1(object):
-    def __init__(self, quark, nucleon, input_dict=None):
+    def __init__(self, quark, nucleon, input_dict):
         """ The nuclear form factor FT1
 
         Return the nuclear form factor FT1
@@ -379,36 +355,30 @@ class FT1(object):
         """
         self.quark = quark
         self.nucleon = nucleon
+        self.ip = input_dict
 
-        if input_dict is None:
-            self.input_dict = Num_input().input_parameters
-            # One should include a warning in case the dictionary
-            # does not contain all necessary keys
-        else:
-            self.input_dict = input_dict
 
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = self.input_dict
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return - ip['mu_at_2GeV'] * ip['BT10up']
+                return - self.ip['mu_at_2GeV'] * self.ip['BT10up']
             if self.quark == 'd':
-                return - ip['md_at_2GeV'] * ip['BT10dp']
+                return - self.ip['md_at_2GeV'] * self.ip['BT10dp']
             if self.quark == 's':
-                return - ip['ms_at_2GeV'] * ip['BT10s']
+                return - self.ip['ms_at_2GeV'] * self.ip['BT10s']
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return - ip['mu_at_2GeV'] * ip['BT10un']
+                return - self.ip['mu_at_2GeV'] * self.ip['BT10un']
             if self.quark == 'd':
-                return - ip['md_at_2GeV'] * ip['BT10dn']
+                return - self.ip['md_at_2GeV'] * self.ip['BT10dn']
             if self.quark == 's':
-                return - ip['ms_at_2GeV'] * ip['BT10s']
+                return - self.ip['ms_at_2GeV'] * self.ip['BT10s']
 
 
 class FG(object):
-    def __init__(self, nucleon, input_dict=None):
+    def __init__(self, nucleon, input_dict):
         """ The nuclear form factor FG
 
         Return the nuclear form factor FG
@@ -421,26 +391,20 @@ class FG(object):
                                  (default is Num_input().input_parameters)
         """
         self.nucleon = nucleon
+        self.ip = input_dict
 
-        if input_dict is None:
-            self.input_dict = Num_input().input_parameters
-            # One should include a warning in case the dictionary
-            # does not contain all necessary keys
-        else:
-            self.input_dict = input_dict
 
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = self.input_dict
 
         if self.nucleon == 'p':
-            return -2*ip['mG']/27
+            return -2*self.ip['mG']/27
         if self.nucleon == 'n':
-            return -2*ip['mG']/27
+            return -2*self.ip['mG']/27
 
 
 class FGtilde(object):
-    def __init__(self, nucleon, input_dict=None):
+    def __init__(self, nucleon, input_dict):
         """ The nuclear form factor FGtilde
 
         Return the nuclear form factor FGtilde
@@ -453,62 +417,57 @@ class FGtilde(object):
                                  (default is Num_input().input_parameters)
         """
         self.nucleon = nucleon
+        self.ip = input_dict
 
-        if input_dict is None:
-            self.input_dict = Num_input().input_parameters
-            # One should include a warning in case the dictionary
-            # does not contain all necessary keys
-        else:
-            self.input_dict = input_dict
 
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = self.input_dict
-        self.mtilde = 1/(1/ip['mu_at_2GeV'] + 1/ip['md_at_2GeV'] + 1/ip['ms_at_2GeV'])
-        self.mN = (ip['mproton'] + ip['mneutron'])/2
+
+        self.mtilde = 1/(1/self.ip['mu_at_2GeV'] + 1/self.ip['md_at_2GeV'] + 1/self.ip['ms_at_2GeV'])
+        self.mN = (self.ip['mproton'] + self.ip['mneutron'])/2
         
         if self.nucleon == 'p':
-            return -self.mN * self.mtilde * (ip['Deltaup']/ip['mu_at_2GeV']\
-                                              + ip['Deltadp']/ip['md_at_2GeV']\
-                                              + ip['Deltas']/ip['ms_at_2GeV'])
+            return -self.mN * self.mtilde * (self.ip['Deltaup']/self.ip['mu_at_2GeV']\
+                                              + self.ip['Deltadp']/self.ip['md_at_2GeV']\
+                                              + self.ip['Deltas']/self.ip['ms_at_2GeV'])
         if self.nucleon == 'n':
-            return -self.mN * self.mtilde * (ip['Deltaun']/ip['mu_at_2GeV']\
-                                              + ip['Deltadn']/ip['md_at_2GeV']\
-                                              + ip['Deltas']/ip['ms_at_2GeV'])
+            return -self.mN * self.mtilde * (self.ip['Deltaun']/self.ip['mu_at_2GeV']\
+                                              + self.ip['Deltadn']/self.ip['md_at_2GeV']\
+                                              + self.ip['Deltas']/self.ip['ms_at_2GeV'])
 
     def value_pion_pole(self):
         """ Return the coefficient of the pion pole
 
         The pion pole is given, in terms of the spatial momentum q, by q^2 / (q^2 + mpi0^2)
         """
-        ip = self.input_dict
-        self.mtilde = 1/(1/ip['mu_at_2GeV'] + 1/ip['md_at_2GeV'] + 1/ip['ms_at_2GeV'])
-        self.mN = (ip['mproton'] + ip['mneutron'])/2
+
+        self.mtilde = 1/(1/self.ip['mu_at_2GeV'] + 1/self.ip['md_at_2GeV'] + 1/self.ip['ms_at_2GeV'])
+        self.mN = (self.ip['mproton'] + self.ip['mneutron'])/2
 
         if self.nucleon == 'p':
-            return self.mN * self.mtilde * ip['gA'] * (1/ip['mu_at_2GeV'] - 1/ip['md_at_2GeV']) / 2
+            return self.mN * self.mtilde * self.ip['gA'] * (1/self.ip['mu_at_2GeV'] - 1/self.ip['md_at_2GeV']) / 2
         if self.nucleon == 'n':
-            return - self.mN * self.mtilde * ip['gA'] * (1/ip['mu_at_2GeV'] - 1/ip['md_at_2GeV']) / 2
+            return - self.mN * self.mtilde * self.ip['gA'] * (1/self.ip['mu_at_2GeV'] - 1/self.ip['md_at_2GeV']) / 2
 
     def value_eta_pole(self):
         """ Return the coefficient of the eta pole
 
         The eta pole is given, in terms of the spatial momentum q, by q^2 / (q^2 + meta^2)
         """
-        ip = self.input_dict
-        self.mtilde = 1/(1/ip['mu_at_2GeV'] + 1/ip['md_at_2GeV'] + 1/ip['ms_at_2GeV'])
-        self.mN = (ip['mproton'] + ip['mneutron'])/2
+
+        self.mtilde = 1/(1/self.ip['mu_at_2GeV'] + 1/self.ip['md_at_2GeV'] + 1/self.ip['ms_at_2GeV'])
+        self.mN = (self.ip['mproton'] + self.ip['mneutron'])/2
 
         if self.nucleon == 'p':
-            return self.mN * self.mtilde * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])\
-                * (1/ip['mu_at_2GeV'] + 1/ip['md_at_2GeV'] - 2/ip['ms_at_2GeV']) / 6
+            return self.mN * self.mtilde * (self.ip['Deltaup'] + self.ip['Deltadp'] - 2*self.ip['Deltas'])\
+                * (1/self.ip['mu_at_2GeV'] + 1/self.ip['md_at_2GeV'] - 2/self.ip['ms_at_2GeV']) / 6
         if self.nucleon == 'n':
-            return self.mN * self.mtilde * (ip['Deltaun'] + ip['Deltadn'] - 2*ip['Deltas'])\
-                * (1/ip['mu_at_2GeV'] + 1/ip['md_at_2GeV'] - 2/ip['ms_at_2GeV']) / 6
+            return self.mN * self.mtilde * (self.ip['Deltaun'] + self.ip['Deltadn'] - 2*self.ip['Deltas'])\
+                * (1/self.ip['mu_at_2GeV'] + 1/self.ip['md_at_2GeV'] - 2/self.ip['ms_at_2GeV']) / 6
 
 
 class FTwist2:
-    def __init__(self, flavor, nucleon, input_dict=None):
+    def __init__(self, flavor, nucleon, input_dict):
         """ The twist-two nuclear form factors
         
         Return the twist-two nuclear form factors
@@ -525,38 +484,32 @@ class FTwist2:
 
         self.flavor = flavor
         self.nucleon = nucleon
+        self.ip = input_dict
 
-        if input_dict is None:
-            self.input_dict = Num_input().input_parameters
-            # One should include a warning in case the dictionary
-            # does not contain all necessary keys
-        else:
-            self.input_dict = input_dict
 
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
 
-        ip = self.input_dict
-        self.mp = ip['mproton']
-        self.mn = ip['mneutron']
+        self.mp = self.ip['mproton']
+        self.mn = self.ip['mneutron']
 
         if self.nucleon == 'p':
             if self.flavor == 'u':
-                return 3/4 * self.mp * ip['f2up']
+                return 3/4 * self.mp * self.ip['f2up']
             if self.flavor == 'd':
-                return 3/4 * self.mp * ip['f2dp']
+                return 3/4 * self.mp * self.ip['f2dp']
             if self.flavor == 's':
-                return 3/4 * self.mp * ip['f2sp']
+                return 3/4 * self.mp * self.ip['f2sp']
             if self.flavor == 'g':
-                return 3/4 * self.mp * ip['f2g']
+                return 3/4 * self.mp * self.ip['f2g']
         if self.nucleon == 'n':
             if self.flavor == 'u':
-                return 3/4 * self.mn * ip['f2un']
+                return 3/4 * self.mn * self.ip['f2un']
             if self.flavor == 'd':
-                return 3/4 * self.mn * ip['f2dn']
+                return 3/4 * self.mn * self.ip['f2dn']
             if self.flavor == 's':
-                return 3/4 * self.mn * ip['f2sn']
+                return 3/4 * self.mn * self.ip['f2sn']
             if self.flavor == 'g':
-                return 3/4 * self.mn * ip['f2g']
+                return 3/4 * self.mn * self.ip['f2g']